% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsc_nlme_stats.R
\docType{data}
\name{gdsc_nlme_stats}
\alias{gdsc_nlme_stats}
\title{The fitted dose response results for gdsc_v17 in a data frame as calculated 
using gdscIC50::calcNlmeStats.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1867107 rows and 26 columns.}
\usage{
gdsc_nlme_stats
}
\description{
This data set id included to aid running of the accompanying vignette.
}
\keyword{datasets}
