% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsc_plate_maps.R
\docType{data}
\name{gdsc_plate_maps}
\alias{gdsc_plate_maps}
\title{Plate maps for the drug sets used in the gdsc_v17 data set}
\format{An object of class \code{data.frame} with 43968 rows and 9 columns.}
\usage{
gdsc_plate_maps
}
\description{
A single data frame containing the well by well treatments fot all drug sets
in the v17 data set. The data is presented in the drug set layout format used
 internally by the GDSC project.
}
\keyword{datasets}
