% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsc_v17.R
\docType{data}
\name{gdsc_v17}
\alias{gdsc_v17}
\title{The Genomics of Drug Sensitivity in Cancer (GDSC)  pharmacogenomics v17 data set}
\format{An object of class \code{data.frame} with 3686353 rows and 16 columns.}
\usage{
gdsc_v17
}
\description{
The GDSC v17 raw data as a single data frame in the format used as input
for dose reponse curve fitting using the gdscIC50 package.
}
\keyword{datasets}
